+++ src/stores/profile.js
@@ -2793,6 +2793,7 @@ export const useProfileStore = defineStore('profile', {
 
         }
         // they changed something
+
         this.dataChanged()
 
       }else{
@@ -2851,8 +2852,6 @@ export const useProfileStore = defineStore('profile', {
                     break
                 }
 
-
-
                 if (!currentUserValuePos[p.propertyURI]){
                     currentUserValuePos[p.propertyURI] = []
                 }
@@ -2863,6 +2862,10 @@ export const useProfileStore = defineStore('profile', {
                   thisLevelType = await utilsRDF.suggestTypeNetwork(p.propertyURI)
                 }
 
+                // if it's a complexSubject, replace bf:Topic with madsrdf:ComplexSubject -- the conversion expects this
+                if (thisLevelType == "http://id.loc.gov/ontologies/bibframe/Topic" && propertyPath.some((obj) => obj.propertyURI == "http://www.loc.gov/mads/rdf/v1#componentList")){
+                  thisLevelType = 'madsrdf:ComplexSubject'
+                }
 
                 let thisLevel = {'@guid':short.generate()}
                 if (!utilsRDF.isUriALiteral(thisLevelType)){
@@ -5733,7 +5736,6 @@ export const useProfileStore = defineStore('profile', {
 
               // if it is an admin metadata do something special
               if (component.propertyURI == "http://id.loc.gov/ontologies/bibframe/adminMetadata"){
-
                 for (let pt in this.activeProfile.rt[component.parentId].pt){
                   if (this.activeProfile.rt[component.parentId].pt[pt].propertyURI == component.propertyURI && this.activeProfile.rt[component.parentId].pt[pt].adminMetadataType && this.activeProfile.rt[component.parentId].pt[pt].adminMetadataType == 'primary' ){
                     ptObjFound = this.activeProfile.rt[component.parentId].pt[pt]
@@ -5768,7 +5770,6 @@ export const useProfileStore = defineStore('profile', {
                         }
                       }
                     }
-
                     console.log("localId",localId)
                     console.log("local040",local040)
 
@@ -5850,23 +5851,36 @@ export const useProfileStore = defineStore('profile', {
                   component = this.componentLibraryUpdateUserValueGuid(component)
 
 
-                  if (Object.keys(ptObjFound.userValue).length <= 1){
+                  if (Object.keys(ptObjFound.userValue).length <= 1 || ptObjFound.id.includes("id_loc_gov_ontologies_bibframe_adminmetadata")){
                     // if this is 1 or 0 then the userdata is empty, with just a @root property
                     // there is no user data added yet
                     // we can just overwrite whats there with our component
                     // we don't need to adjust the order, its 1-for-1
                     // find it again and overwrite
-                    for (let pt in this.activeProfile.rt[component.parentId].pt){
-                      if (this.activeProfile.rt[component.parentId].pt[pt].id == component.id){
-                        this.activeProfile.rt[component.parentId].pt[pt] = JSON.parse(JSON.stringify(component))
-                        this.dataChanged()
-                        return [component.parentId,pt]
+
+                    // Admin metadata should overwrite the existing values, otherwise we create a new Admin field, which we shouldn't do
+                    if (!ptObjFound.id.includes("id_loc_gov_ontologies_bibframe_adminmetadata")){
+                      for (let pt in this.activeProfile.rt[component.parentId].pt){
+                        if (this.activeProfile.rt[component.parentId].pt[pt].id == component.id){
+                          this.activeProfile.rt[component.parentId].pt[pt] = JSON.parse(JSON.stringify(component))
+                          this.dataChanged()
+                          return [component.parentId,pt]
+                        }
+                      }
+                    } else {
+                      // Need an exception for Admin to ensure we're not making changes to a different admin field
+                      for (let pt in this.activeProfile.rt[component.parentId].pt){
+                        if (this.activeProfile.rt[component.parentId].pt[pt].id == ptObjFound.id){
+                          component.id = ptObjFound.id
+                          this.activeProfile.rt[component.parentId].pt[pt] = JSON.parse(JSON.stringify(component))
+                          this.dataChanged()
+                          return [component.parentId,pt]
+                        }
                       }
                     }
 
 
                   }else{
-
                     // we can't replace the one that is there, already has data, so construct a new place for it
 
                     // first find out how many of these components there are
@@ -5894,7 +5908,7 @@ export const useProfileStore = defineStore('profile', {
                       }
                     }
                     this.activeProfile.rt[component.parentId].ptOrder.splice(insertAt+1, 0, newId);
-
+                    this.dataChanged()
                     return [component.parentId,newId]
                   }